% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdr_enroll_MDCR_ENROLL_AB_01.R
\docType{data}
\name{MDCR_ENROLL_AB_01}
\alias{MDCR_ENROLL_AB_01}
\title{Total Medicare Enrollment}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 9 columns.}
\usage{
MDCR_ENROLL_AB_01
}
\description{
NOTES:  The enrollment counts are determined using a person-year methodology.  Numbers and percentages may not add to totals because of rounding.
SOURCE:  Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, CMS Chronic Conditions Data Warehouse.
}
\keyword{datasets}
