% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR_ENROLL_AB_06.R
\docType{data}
\name{MDCR_ENROLL_AB_06}
\alias{MDCR_ENROLL_AB_06}
\title{Total Medicare Enrollment:  Part A and/or Part B Enrollees, by Type of Entitlement and Demographic Characteristics}
\format{An object of class \code{data.frame} with 110 rows and 10 columns.}
\source{
Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, CMS Chronic Conditions Data Warehouse.
}
\usage{
MDCR_ENROLL_AB_06
}
\description{
Total Medicare Enrollment:  Part A and/or Part B Enrollees, by Type of Entitlement and Demographic Characteristics
}
\details{
NOTES:  The enrollment counts are determined using a person-year methodology.  Numbers and percentages may not add to totals because of rounding.
}
\keyword{datasets}
