% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR_ENROLL_AB_07.R
\docType{data}
\name{MDCR_ENROLL_AB_07}
\alias{MDCR_ENROLL_AB_07}
\title{Total Medicare Enrollment:  Part A and/or Part B Total, Aged, and Disabled Enrollees, by Area of Residence}
\format{An object of class \code{data.frame} with 300 rows and 14 columns.}
\source{
Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, CMS Chronic Conditions Data Warehouse.
}
\usage{
MDCR_ENROLL_AB_07
}
\description{
Total Medicare Enrollment:  Part A and/or Part B Total, Aged, and Disabled Enrollees, by Area of Residence
}
\details{
NOTES:  The enrollment counts are determined using a person-year methodology.  Numbers and percentages may not add to totals because of rounding.
}
\keyword{datasets}
