% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR_ENROLL_AB_08.R
\docType{data}
\name{MDCR_ENROLL_AB_08}
\alias{MDCR_ENROLL_AB_08}
\title{Total Medicare Enrollment:  Part A and/or Part B Enrollees, by Type of Entitlement and Area of Residence.}
\format{An object of class \code{data.frame} with 300 rows and 9 columns.}
\source{
Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, CMS Chronic Conditions Data Warehouse.
}
\usage{
MDCR_ENROLL_AB_08
}
\description{
Total Medicare Enrollment:  Part A and/or Part B Enrollees, by Type of Entitlement and Area of Residence.
}
\details{
NOTES:  The enrollment counts are determined using a person-year methodology.  Numbers and percentages may not add to totals because of rounding.

There are some NA values in this data set.  These NA values are associated with counts between 1 and 10 have been suppressed because of CMS rules to protect the privacy of beneficiaries, or have been cross-suppressed to prevent the recalculation of suppressed counts between 1 and 10.
}
\seealso{
\code{vignette("program-statistics", package = "cms.program.statistics")}
}
\keyword{datasets}
