% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLACE_OF_SERVICE.R
\docType{data}
\name{PLACE_OF_SERVICE}
\alias{PLACE_OF_SERVICE}
\title{Place of Serivce}
\format{An object of class \code{data.frame} with 99 rows and 3 columns.}
\usage{
PLACE_OF_SERVICE
}
\description{
Place of Serivce
}
\keyword{datasets}
