% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_taxonomy.R
\docType{data}
\name{provider_taxonomy}
\alias{provider_taxonomy}
\title{Provider Taxonomy}
\format{An object of class \code{data.frame} with 507 rows and 4 columns.}
\usage{
provider_taxonomy
}
\description{
Important Notes:
}
\details{
Medicare no longer enrolls the provider type: Voluntary Health or Charitable Agency

"Mass Immunizer Roster Biller" is an arrangement whereby immunization services are rendered by health care providers who are under contract to the entity/individual who is enrolled in Medicare as a Mass Immunization Roster Biller. The Healthcare Provider Taxonomy Code(s) would be that (those) of the health care provider(s) who render the immunization services.

Medicare carriers use Specialty Code 87 for Pharmacies when processing prescription drugs that are paid through the Medicare Physician Fee Schedule.  However, the description for Specialty Code 87 remains "All Other Suppliers."

Because the Medicare Provider Type "Physician/Undefined Physician Type" applies to physicians, "Unknown Supplier/Provider Specialty" applies to other than physicians.  Because Medicare does not enroll every non-physician supplier/provider type in the Healthcare Provider Taxonomy Code Set, we cannot arbitrarily link this Medicare Provider Type to all remaining supplier/provider types in the Healthcare Provider Taxonomy Code Set.  In truth, this Medicare Provider Type would link to every Healthcare Provider Taxonomy Code (except those that describe physicians) that is listed in this document.  We do not list them here because doing so would greatly increase the size of this document.

Because the Medicare Provider Type 96 indicates that the provider/supplier is a Physician, we have linked this to the most generic All opathic & Osteopathic Physicians classification.  It could technically link to every Allopathic & Osteopathic Physician.

Medicare specialty code designation "Physician/Undefined Physician type" is for Part B DMEPOS purposes.

Medicare specialty code designation "A0" is for Part B DMEPOS purposes.

Medicare specialty code designation "A1" is for Part B DMEPOS purposes.

Medicare specialty code designation "A2" is for Part B DMEPOS purposes.

Medicare specialty code designation "A3" is for Part B DMEPOS purposes.

Medicare specialty code designation "A4" is for Part B DMEPOS purposes.

RE: "A9": There is no linkage to specific Healthcare Provider Taxonomy Codes because an Indian Health Service facility can be various provider types/classifications.

Medicare specialty code designation "Indian Health Service facility" is for Part B DMEPOS purposes.

RE: "B4": Does not meet the definition of "health care provider" at 45 CFR 160.103; is not eligible for an NPI.
}
\keyword{datasets}
