% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR_ENROLL_AB_01.R
\docType{data}
\name{MDCR_ENROLL_AB_01}
\alias{MDCR_ENROLL_AB_01}
\title{Total Medicare Enrollment:  Total, Original Medicare, and Medicare Advantage and Other Health Plan Enrollment}
\format{An object of class \code{data.frame} with 10 rows and 9 columns.}
\source{
Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, CMS Chronic Conditions Data Warehouse.
}
\usage{
MDCR_ENROLL_AB_01
}
\description{
Total Medicare Enrollment:  Total, Original Medicare, and Medicare Advantage and Other Health Plan Enrollment
}
\details{
NOTES:  The enrollment counts are determined using a person-year methodology.  Numbers and percentages may not add to totals because of rounding.
}
\keyword{datasets}
