% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDCR_ENROLL_AB_02.R
\docType{data}
\name{MDCR_ENROLL_AB_02}
\alias{MDCR_ENROLL_AB_02}
\title{MDCR ENROLL AB 02}
\format{An object of class \code{data.frame} with 255 rows and 12 columns.}
\usage{
MDCR_ENROLL_AB_02
}
\description{
Total Medicare Enrollment:  Total, Original Medicare, Medicare Advantage and Other Health Plan Enrollment, and Resident Population, by Area of Residence
}
\details{
Counts between 1 and 10 have been suppressed because of CMS rules to protect the privacy of beneficiaries.

United States Totals Excludes territories, possessions, foreign countries, and unknown.

Total United States and state population estimates are available on the U.S. Census Bureau's website:  http://www.census.gov/popest/

NOTES:  The enrollment counts are determined  using a person-year methodology.  Numbers and percentages may not add to totals because of
rounding.  For definitions of "Metropolitan", "Micropolitan", and "Non-Core-Based Statistical Area", refer to CPS Glossary.

SOURCES:  Centers for Medicare & Medicaid Services, Office of Enterprise Data and Analytics, Chronic Conditions Data Warehouse; United States Census Bureau.
}
\keyword{datasets}
